refactor(tasks): Rename processors package to tasks
- Move all modules from `ampf/processors` to `ampf/tasks`
- Update all internal and external imports to reflect the new package name
- Update `README.md` and documentation to reference the `tasks` package
- Improve clarity and consistency by using a more descriptive package name
