+++ README.md
@@ -15,6 +15,7 @@ Set of helper classes:
   * [BaseDecorator](doc/base_decorator.md) - simple class to create **Decorator** pattern.
   * [SubscriptionProcessor](doc/base_subscription_processor.md) - base class for processing messages from a subscription.
 * [Dependency](doc/dependency.md) - simple dependency registry for managing dependencies in your application.
+* [Tasks](doc/tasks.md) - helper for running background tasks
 * FastAPI - helper classes for FastAPI framework
   * [StaticFileResponse](doc/fastapi/static_file_response.md) - return static files or index.html (Angular files)
   * [JsonStreamingResponse](doc/fastapi/json_streaming_response.md) - streams Pydantic objects to client as JSON.
