+++ ampf/tasks/direct_runner.py
@@ -2,13 +2,11 @@ from typing import override
 
 from pydantic import BaseModel
 
-from ampf.processors.task_registry import TaskRegistry
-
 from .task_model import TaskRunner
+from .task_registry import TaskRegistry
 
 
 class DirectRunner(TaskRunner):
-    
     @override
     def run(self, name: str, payload: BaseModel):
         TaskRegistry.run_task(name, payload)
